// backend/src/repository/enums/availability.enum.ts
export enum Availability {
  ELIGIBLE = "ELIGIBLE",
  PREFERENCE_COURSE = "PREFERENCE_COURSE",
  OVERTIME_FULL = "OVERTIME_FULL",
  OVERTIME_PART = "OVERTIME_PART",
  OUT_OF_CAPABILITY_TIME = "OUT_OF_CAPABILITY_TIME",
  OUT_OF_CAPABILITY_DAY = "OUT_OF_CAPABILITY_DAY",
  OUT_OF_CAPABILITY_COURSE = "OUT_OF_CAPABILITY_COURSE",
  IN_DAYOFF = "IN_DAYOFF",
  OVERLAPPING_SCHEDULE = "OVERLAPPING_SCHEDULE",
}